better default arguments via bare() objects
